.profile: fix spawn_ssh_agent
I recently had to deal with a system that doesn't have systemd (Alpine),
and it used this "deprecated" way of launching ssh-agent.
It didn't work with agent forwarding though: the local agent would be
launched regardless, somehow overriding the forwarded agent.
This is fixed by testing if SSH_AUTH_SOCK (instead of SSH_AGENT_PID) is
defined, and it is defined for agent-forwarded ssh sessions.

It remains to be seen how this interacts with systems that use the
systemd unit to launch ssh-agent.